# Revamp AES Part 1; including random & encrypted IV, hashed key and max keysize