--- a/benchmark.c
+++ b/benchmark.c
@@ -35,7 +35,7 @@ static int gettimeofday(struct timeval *tp, void *tzp)
 	time_t clock;
 	struct tm tm;
 	SYSTEMTIME wtm;
-	GetLocalTime(&wtm);
+	GetSystemTime(&wtm);
 	tm.tm_year = wtm.wYear - 1900;
 	tm.tm_mon = wtm.wMonth - 1;
 	tm.tm_mday = wtm.wDay;