This pull request addresses some security issues in the AES module _transform_aes_ as follows:

**Key Stretching**
- the key provided by the user gets hashed
- this hash is used to derive the actual encryption and decryption key
- also, a secondary key is derived from the hash
- for being able to derive two independent keys in the same step, a hash function with wider output is used (SHA512)

**CBC data encryption**
- make use of an full block sized initialization vector (IV) instead of a nonce
- the IV is randomly set for every packet
- as the first data block is quite predictable (preamble, NIC's MACs), the IV gets encrypted as single block using the secondary key
- replay protection could be added to the IV and may be considered in Part 2

**Message authentication**
- a field for a message authentication code is filled with a ECBC-MAC, i.e. the last cipher block gets encrypted using the secondary key
- to ensure safety of ECBC-MAC, i.e. keep it prefix-free, data length is prepended to the data to be encrypted
- thus, the number of padded bytes at the last packet's end is not neccessary anymore and gets removed
- verification of the message authentication code was not first priority and may be included in Part 2

**Any other business**
- version number increased to 2 as packet design changed
- fixes a packet loss issue - at least for local edges
- author's first pull request ever - please be kind :)