@emanuele-f 2019-03-03
Thank you! The pull request is very clean, your contribution is very welcome. I'm adding some inline comments for the review. If possible, also provide some links to online resources which explain the concepts used, for example the "prefix free" part of ECBC-MAC. Here is a reminder for me of the encryption/decryption process (please correct me if something is wrong).

```
*Encryption*:
iv = randomIv()
aes_cbc_encrypt(iv, raw_packet, primary_key) -> encryped_payload
aes_encrypt(iv, secondary_key) -> encrypted_iv
aes_encrypt(last_block, secondary_key) -> message_mac

*Decryption*:
aes_decrypt(encrypted_iv, secondary_key) -> iv
aes_cbc_decrypt(iv, encryped_payload, primary_key) -> raw_packet

*Authentication* (todo):
aes_decrypt(last_encrypted_block, secondary_key) -> transmitted_mac
aes_encrypt(last_encrypted_block, secondary_key) -> expected_mac

if transmitted_mac != expected_mac
  drop_packet()
```

Please also note that this adds a 34 byte overhead to every packet (reducing this may be desiderable). A comparison of the results of the `benchmark` program available in n2n would also be useful to understand the performance impact of the changes.