@Logan008 2019-03-03
Thank you for the feedback! I give additional comments as you suggested in reply to your inline comments above.

Still learning git, I wonder how I can make the additional changes to the code as suggested (your comments are very valuable)? New Pull Request? Please advice.