@emanuele-f 2019-03-03
Thank you, I've updated the comment above for the authentication part. In order to update this pull request you only need to commit the new changes to your branch. So, after making the changes, do a `git add ...your_changed_files...`, `git commit` and then push the changes to your online branch with `git push`. This will automatically update the pull request, since your online branch is now *bound* to this pull request