@Logan008 2019-03-04
I have just finished implementing the changes you suggested and applied some obvious corrective changes.

You are right, concerning packet size, we now see longer packets: I would just put "minus 1" for dropping the padding length. However, to keep it a little bit more flexible, I implemented the message authentication in a way that allows to change its length in bytes at compile time - just change the definition (0 ... 16, all seem to work well). As the length of the message authentication code is not explicitly transmitted, a change in message authentication code size causes incompatibility to clients with a different size. Maybe something to ponder for Part 2...

For now, there is just the need to agree on a reasonable value for it.