@Logan008 2019-03-07
The aes_version and sa number now are included in the message authentication by just xor'ing them before the message authentication code gets encrypted using K2. So, the accordingly updated message authentication verification scheme should look like this:

```
**transmitted_mac** from "header" in front of IV

aes_cbc_encrypt(decrypted_packet, key) to another buffer
-> xor'ing in the aes_version and sa field from the received packet to the last block
-> aes_encrypt (last_block_of that, secondary key) 
-> **expected_mac**

if transmitted_mac != expected_mac
  drop_packet()
```

Leftovers for Part 2 of AES revamp: Encounter Replay Attacks by modifying IV scheme (could remain compatible to this aes version) and clean up of sa structure.

I did some benchmarking (i7-5775C). The modifications indeed cost a bit of performance as follows:

Before modifcation (6.5 x slower than null):
```
./benchmark 
..........
[transop_null] 10000 times: (5255 -> 525 nsec each) 1551993236.392130 -> 1551993236.397385.
..........
[transop_twofish] 10000 times: (269364 -> 26936 nsec each) 1551993236.397415 -> 1551993236.666779.
..........
[transop_aes] 10000 times: (34192 -> 3419 nsec each) 1551993236.666785 -> 1551993236.700977.
```

After modification (15 x slower than null)
```./benchmark 
..........
[transop_null] 10000 times: (4033 -> 403 nsec each) 1551993168.241907 -> 1551993168.245940.
..........
[transop_twofish] 10000 times: (267498 -> 26749 nsec each) 1551993168.245970 -> 1551993168.513468.
..........
[transop_aes] 10000 times: (62269 -> 6226 nsec each) 1551993168.513472 -> 1551993168.575741.


```

So the gain of AES's hardware acceleration is melting a bit, but it is still way ahead of TF. Possible shortcuts could be taken by later maybe adding a command line parameters for fixed, lower keysize and disabled message authentication. Though, I would not recommend that at all from a security point of view.