@Logan008 2019-03-11
Sorry, if this pull request now gets a little bit lengthy... I have just added replay protection by encoding a  time stamp to the initialization vector as outlined in #73.

There are two values to parametrize the replay protection (definitions at the top of transform_aes.c):

- One is the time frame in which packets with a certain time stamp are accepted by the receiver by comparing the timestamp applied by the sender to the receiver's system time. This is set to +/- 16 seconds, allowing for 32 seconds of time difference between the two clocks plus transmission time. In the age of ubiquitious NTP servers, this might be reasonable.
- Nodes keep track of the previous packet's time stamp and make sure, that only packets with higher time stamp are accepted - assuming that packets do not overtake each other on the line too often. But as exactly this seems to happen from time to time, there is a grace time set to 160 ms which seems to work well in most cases, at least in the author's setup.

Also, I changed the message authentication code length to 4 bytes. Would that be a reasonable value?

The replay prevention does not come for free: There is a slight performance slump as follows (15.8 x slower than null, see the post above):

```
./benchmark
..........
[transop_null] 10000 times: (3732 -> 373 nsec each) 1552294709.652039 -> 1552294709.655771.
..........
[transop_twofish] 10000 times: (268661 -> 26866 nsec each) 1552294709.655796 -> 1552294709.924457.
..........
[transop_aes] 10000 times: (74913 -> 7491 nsec each) 1552294709.924461 -> 1552294709.999374.
```

By the way, talking performance, I found that key size has a somewhat non-neglectable impact on performance. Changing the now hard-wired keysize from 256 to 128 bit makes it look as follows (1400 ns faster):

```
./benchmark
..........
[transop_null] 10000 times: (3230 -> 323 nsec each) 1552293781.135773 -> 1552293781.139003.
..........
[transop_twofish] 10000 times: (270719 -> 27071 nsec each) 1552293781.139030 -> 1552293781.409749.
..........
[transop_aes] 10000 times: (60573 -> 6057 nsec each) 1552293781.409753 -> 1552293781.470326.
```

AES version incremented to 3. **This shall conclude the changes I wanted to apply for now.** Please let me know what you think and please test. Do the times for replay attack protection work for you? What does it take to make this pull request acceptable? I would love to hear back from you to make n2n even better.

A next step (new pull request, later) could include implementation of support for aes version 2 by iffing out the message authentication, the replay attack protection and also maybe switching to a lower keysize - allowing for reasonable use on less performing systems, e.g. routers.