@Logan008 2019-03-16
Author's evaluation after five days: works flawless in a setup with one public supernode and four edges around the world.